services/worker-service/pyproject.toml
[project]
name = "worker-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.13.0",
    "asyncpg>=0.31.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]
